fix(configuration): rename refactor and update demo page
