+++ packages/elements/src/configuration/index.ts
@@ -1,2 +1,2 @@
-export * from './elements/config.js';
+export * from './elements/configuration.js';
 export { efConfig, type Config } from './helpers/context.js';
